Github issue: Fix IAuthService issue when returning Results.Forbid() #1
The /increment endpoint failed when the INCREMENT feature flag was disabled "INCREMENT=false" , the fix explicitly defines all possible responses either Ok response as expected when the feature flag is disabled and 403 "forbidden" when it is enable.

To test this run the docker compose file and just make a post request to /increment and change the env INCREMENT to either "false" or "true" (careful with spaces needs to be clean) and you will see the desired outputs.